Create Autonomous_Selection.c
Autonomous_Selection.c

In the VEX competition, it is almost necessary to prepare multiple autonomous routines, as the robot might start the match in one of four positions that won’t be known until shortly before the match.  There is also strategic value to having multiple plans of attack from each starting point, depending on the opponents.  This file contains all the code to allow a user to select an autonomous routine at the beginning of a match.  The left and right LCD buttons scroll through the available options, displayed on the screen, and the center button enters the selection.