Create Autonomous_Modes.c
Autonomous_Modes.c

This file contains the autonomous routines themselves.  They are highly flexible, being built from the autonomous functions established in “Subsystem_Functions.c”, and can be customized to suit different scenarios.  For safety, the default option is to do nothing.  This requires a user to intentionally select a routine if the robot is to move at all.  There are separate routines for each starting position, and more can be added if necessary.